Make use of jlm::util::filepath where applicable (#758)
In a couple of places `std::string` was used for file paths, which I
changed.

Makes the constructor of `jlm::util::filepath` explicit, which exposed a
few places where implicit conversions were used.

In order to make it more ergonomic to use `filepath`, an overload for
`Join` taking a string was added, and `WithSuffix` was added.

`filepath` also got a helper for getting the path to the temp directory.