add HashSet::DifferenceWith method (#790)
I had a use for this method in the Andersen-work, and thought I would
backport it as I am likely to want to use it again.

With this operation, there are two possible implementations. I try to
pick the one that should be the fastest, but I have not done
benchmarking to figure out exactly where one implementation becomes
faster than the other.