Add support for IOBarrier [AndersenAgnostic][SteensgaardAgnostic] (#795)
This PR introduces the IOBarrier operation and adds support for it to
the LLVM and HLS dialect. The IOBarrier operation is necessary to
sequentialize operations with potentially undefined behavior, such as
division and modulo by zero as well as load and store operations from
null pointers, to ensure correct program semantics. For example, the
following code requires a IOBarrier to avoid unwanted undefined
behavior:

```
int f(int x, int y)
{
   opaque(); // calls internally exit(0)
   return x / y;
}
```
The program is a valid C program under all circumstances, even if y
equals zero. In order to capture this semantic, it is necessary to
sequentialize the division operation below the call to opaque in the
RVSDG. This is accomplished using the IOBarrier operation. If it can be
proven that an operation can never exhibit undefined behavior, then the
respective IOBarrier can be removed.

Close #750