Add LLVM-specific integer comparison operations (#801)
This PR replaces the old bitstring comparison operations with new LLVM
specific
integer comparison operations. The implementations have currently still
the following limitations:

1. The operations still use the bistring type. The type will potentially
be
later replaced by an LLVM specific type.
2. No operation specific reductions have been implemented yet. This will
also happen in a follow-up PR.

At this point, the implementation only tries to get on par with the
respective bitstring operations (minus the reductions). More LLVM
specific semantics are going to be introduced in a future PR.

---------

Co-authored-by: HKrogstie <krogstie.havard@gmail.com>