Give the cfg_node iterators for outedges and inedges (#802)
While doing other things, I was annoyed with the iterator interfaces for
accessing in-edges and out-edges on cfg nodes not implementing
everything needed to be regarded as real iterators.

I added a helper class for wrapping iterators to pointers, to become an
iterator to the pointees instead. It works for both raw pointers and
smart pointers. See the test for examples.

I also removed some unnecessary `inline` keywords when I was touching
functions.