Give the hls workflow permission to comment on the PR (#822)
I am not entirely sure why, but the HLS workflow on #803 fails when it
tries to write a comment about changed cycle count.

This has never been an issue before, but it might be because the creator
of the PR, or the last committer, is @halvorlinder who has fewer
permissions.

I am hoping this change might help with this, but the permission
documentation is not very clear