Allow the HLS cycle count changed comment on PRs to fail (#825)
It turns out that specifying permissions directly did not help.
The `GITHUB_TOKEN` created for a `pull_request` workflow only has
permissions inside the source repository. Since @halvorlinder's PR #803
is from his own fork, the workflow can not comment on the PR in this
repository.

I found one workaround on [stack
overflow](https://stackoverflow.com/a/71683208) that uploads the comment
as an artifact, and uses a second workflow triggered by the completion
of the first workflow to comment.

However, for an informative comment like this, I think it makes more
sense to just allow the commenting to fail silently.