[AndersenAgnostic] Fix RegionAwareMemoryNodeProvider, now based on call graph and handling calls to external (#827)
Introduces a call graph creation step to the RegionAwareMemoryProvider,
which ensures that all possible call paths are handled, even when going
via external function calls.

SCCs are found in this call graph, and they are visited in reverse
topological order.
This means that for any function call found during the "annotation"
step, it either points to a function in an earlier SCC which is already
finished, or in the same SCC. In the latter case a flag is used to
remember that the call may be recursive.

RegionSummaries also remember if they may contain recursion. Once all
functions have been annotated, a final pass goes through all summaries
that may contain recursion, and add the whole SCC's summary of memory
locations to it.

These changes made the encoding more precise in some of the tests, so
they have been updated to account for that.

Testing the llvm-test-suite with `--AAAndersenRegionAware` passes on all
tests, though I managed to get the encoder to fail when combining
`--AAAndersenRegionAware` with other passes. I felt this PR was large
enough already, so fixing encoder issues is not done here.

I am adding the [AndersenAgnostic] on this PR, but that is only because
[AndersenRegionAware] does not exist yet.