Script for roundtipping via MLIR output / input
The script acts as a jlc replacement by passing all provided
arguments to jlc and appends '-###'.
All found jlm-opt commands are converted to two commands with the
first command consuming the llvm input and generates an mlir output
and the second command consumes the generate mlir output and
generates the llvm output.
All other commands are executed as stated by 'jlc -###'.

It also adds an MLIR roundtrip of polybench to the CI.