Unify output handling in DNE mark phase (#894)
Unifies the output handling in DNE's mark phase. This PR is in
preparation for refactoring DNE such that it is possible to inject
handlers for different kind of nodes. This allows DNE to become
extensible for different node types and ultimately would enable us to
unify the two DNE transformations in the code base (LLVM and HLS).