GammaNode: Add pattern matching support to gamma
Add an argument to gamma branch subregions that represents the "content" of
the discriminator object that gamma is matching over. For booleans/control
decisions, they do not have a content (although one could argue that they
carry an "assertion" that a certain proposition holds that in turn some
operations in the contained region may be dependent on -- e.g. a divisor
not being zero, a pointer being non-null etc.) -- so use UnitType here
by default.

This allows Gamma to serve as general pattern matching operator which makes
the contents of e.g. an algebraic type being matched over available in
its sub branches.

Adapt all places using Gamma API to deal with the fact that there is another
argument now not corresponding to any EntryVar.