Introduce UnitType (#897)
Introduce the UnitType to allow representing a type that holds no data
(akin to the C "void" type). This can formally be used in places where
"something" must be produced, but it has no content -- this will later
be used in pattern matching over types without a content in generalized
gamma constructs.

Prepare places that currently balk on existency of this type so its
usage can be introduced without breakage later.