GammaNode: EntryVar/ExitVar removal API, and consistent usage of safe… (#898)
Add methods that allow safe removal of entryvars / exitvars, and
generally make use of the safe EntryVar/ExitVar API in all places that
need to map between inputs/arguments and results/outputs. This is in
preparation to changes how gamma arguments are allocated when
generalizing gamma to support pattern matching.