Distinguish between node type and operation in region contents check
Debug code uses Region::Contains as a test to check for certain nodes.
Currently it can only check on "Operation", but it code also wants to
check on structural nodes. When reworking operations, that distinction
needs to be accounted for in the code.

Eventually it might make sense to provide a "predicate"-based search API
instead that allows to both check for presence and get access to nodes
of interest.