Separate SimpleOperation / StructuralOperation
Separate the inheritance hierarchies of SimpleOperation / StructuralOperation.
This makes "SimpleOperation" the root class for all operational nodes.
The inheritance hierarchy for StructuralOperation is kept for the moment,
eventually the "contents" of structural operations can be entirely
disassociated. They should evolve into something that can carry information
specific to the structure, particularly backend-specific customizations
of structural nodes.