Fix bug in InvariantValueRedirection pass (#976)
The pass falsely assumed that if the CallMemoryStateExitSplit node is
present for a call, then all the other memory state nodes (i.e.,
CallMemoryStateEntryMerge, LambdaMemoryStateEntrySplit,
LambdaMemoryStateExitMerge) are present as well. This does not have to
be the case in general. For example, it might be that after region-aware
state encoding, the LambdaMemoryStateEntrySplit node is missing, while
all other nodes are present.

Close #937

---------

Co-authored-by: Magnus Sjalander <work@sjalander.com>
Co-authored-by: HKrogstie <krogstie.havard@gmail.com>