Add the option of handling unification during TarjanSCC. (#990)
The Topological Worklist Policy and Online Cycle Detector has not been
handling unification properly due to the TarjanSCC algorithm not
handling unification. This PR adds the ability for Tarjan to handle
graphs with node unification.

As a consequence, it became more practical to provide the topological
order in reverse, which is why some other functions have been changed