Add PointsToGraphAliasAnalysis and tests (#1164)
It also includes the `ChainedAliasAnalysis` because it is a very small
class.

I also had to add some `#includes` to make things compile on my machine
after a system update.

`TryGetConstantSignedInteger` is moved to `AliasAnalysis.{cpp,hpp}`
because it is needed by both analyses, but ideally it should live
somewhere even more general. It uses `NormalizeOutputs` however, which
should also be moved somewhere more general.