Add Alias Analysis Precision Evaluator (#1211)
This is the PrecisionEvaluator for providing statistics about a given
AliasAnalysis and its responses.

It also contains configuration options for dumping precision metrics per
function to a file, as well as dumping a DotWriter graph that contains
nodes for pointers, and edges representing MustAlias and
MayAlias-relations between nodes.

This code should probably have some tests, as there are some moving
parts regarding de-duplication of pointers. Even when pointer
deduplication is not requested, it still aggregates pointers to make the
runtime not explode, but tracks the aggregated values such that the
final results are not affected.