Add statistics about memory states to the MemoryStateEncoder (#1216)
This PR counts the number of memory states, and what type of
`MemoryNode` they correspond to, for the following:
 - Region arguments (Intra-procedural only)
 - Loads (including volatile)
 - Stores (including volatile)
 - Call entry merges

It also tracks the largest amount of memory states used by a single such
entity.

Finally, it tracks how many of the memory states correspond to
`MemoryNode`s that are marked as module escaping.

I should note that adding this to the `MemoryStateEncoder` was done
because it is the last step where we still have access to the underlying
`MemoryNode`s