Replace Value/StateType as classes, implement as traits
Remove the ValueType and StateType as base classes for further types --
every type used in rvsdg can just inherit from Type instead. Indicate
whether a type is of "Value" or "State" kind by an attribute instead.

This is based on the following observations:

- there are only very few places where this distinction really matters

- where it matters (e.g. record elements), "ValueType" is actually not
  sufficiently restrictive

In summary, the distinction gains very little, and where it matters is
better conveyed through more flexible attribute semantics.