HLS: Fix pass-through loop variable removal in UnusedStateRemoval pass (#1244)
The removal of pass-through loop variables in the UnusedStateRemoval
pass was simply broken. It removed happily loop variables even if they
were not just simple pass-through. If a loop variables post value
operand was simply another loop variables pre operand, then the loop
variables were already removed. This is simply wrong. The unit test
exemplifies this with loop variable Y and Z, which take each others
value.

This PR solves an issue that arose in PR #1242