Change traversers and remove trackers (#1221)
Remove dependence on tracker and node depth for both top-down and
bottom-up traversers. Replace it with normal BFS graph traversal
instead.

Trackers are not needed anymore anywhere. Remove them as they rely on
node depth (which is also supposed to go away).

Traversers are fully deterministic and particularly independent of any
addresses; it depends only:
- the order in which nodes are registered as "top" or "bottom" nodes in
graph
- the order in which successor or predecessor nodes are attached to
outputs and inputs

---------

Co-authored-by: Magnus Sjalander <magnus.sjalander@ntnu.no>
Co-authored-by: Magnus Sjalander <work@sjalander.com>