Add scalar evolution pass with method for finding induction variables (#1220)
This PR adds a simple barebones ScalarEvolution class as well as an
optimization command to run it.

Right now, all it does is detect the simplest cases of induction
variables in theta-nodes. A (basic) induction variable is a variable
which changes with a constant value each iteration.

It uses a fixed-point analysis with a bottom-up approach where it starts
with all loop variables in a set, and systematically removes nodes that
are not induction variables. At the end, only the induction variables
remain in the set.

To run the pass, use 
```bash
$ jlm-opt --ScalarEvolution -o test test.ll
```

---------

Co-authored-by: Nico Reissmann <nico.reissmann@gmail.com>