Fix MemoryStateEncoder not creating MemoryStateJoin operations for allocas (#1296)
If the memory state for an alloca is routed into a region, it must be
joined with the alloca.
This naturally involves some changed tests, which now need to account
for extra join nodes.

This PR also changes the encoding of the `malloc` operation to use
`MemoryStateJoin` instead of `MemoryStateMerge`.