Remove Cache and Unknown memory from MemoryStateEncodeer (#1297)
This is a step towards #1288, which changes how the `MemoryStateEncoder`
works internally to make it more ready for the new `ModRefSummary`
interface.

The changes include:
 - Remove the MemoryNodeCache
 - Remove tracking of states related to Unknown memory

It has its own shim between the old interface and the new one, called
`GetSimpleNodeModRef`, which looks a bit ugly now, but will be greatly
simplified with #1288. The `ReplacementMap_` will also be removed.

Lastly, I discovered a bug in the memcpy encoding when updating the
interface. The current implementation will add duplicate inputs and
output if the same state is targeted by both the `src` and `dst` part of
the memcpy:
<img width="467" height="407" alt="image"
src="https://github.com/user-attachments/assets/618da2ef-0437-4f33-ac9f-ec687dbdf15f"
/>

This PR accidentally fixed this, so that is why some tests are touched.