Rework traversers and notifiers (#1299)
Introduce RegionObserver, use it in traversers

This introduces the concept of a "RegionObserver" which -- when alive --
can observe property changes of a region (node addition / removal and
edge changes). Use this instead of global notifier chains for
traversers.

Since notifications are now per-region, everything is managed through an
object attached to a region. Using a singular RegionObserver object
instead of individual callbacks allows to limit the growth of the
"Region" object to a single pointer.

This has the following knock-on effects:
- traversers become smaller, and need to manage a bit less (simplified
observer setup, no "region" object anymore)
- since notification is per region, this avoids quadratic complexity for
nested regions/traversers
- can delete entirety of callback / notifier related code

While at it, the "revisit" feature of BottomUpTraverser was not used
anymore, and was in fact broken. So removing this as well.