Harden API for adding node inputs/outputs, and notify region about new/removed inputs (#1310)
Makes the `addInput`, `addOutput`, `removeInput` and `removeOutput`
methods on `Node` protected, such that subclasses of `Node` have the
full control over the creation of inputs and outputs, and can maintain
the node looking like expected.

This required some changes in passes that directly operated on
inputs/outputs/arguments/results, mostly in HLS.
The HLS `LoopNode` gets some new methods for creating the different
kinds of inputs:
  - loop variable
  - loop constant
  - memory request response

The `TestStructuralNode` gets methods to give tests full freedom to do
whatever with inputs and outputs.

The `addInput` and `removeInput` methods gets an additional `bool
notifyRegion`. When the node creates inputs in its constructor, this
should be false. When creating inputs outside of the constructor it
should be true. Same applies to removing inputs and the destructor.

The logic here is that it would be silly to call `notifyInputCreated`
for every input we create, when the region has not even been notified
about the node being created in the first place.

Adding an extra `bool` parameter to `addInput` should have a quite
limited blast radius, since the PR locks down the `addInput` API.

Lastly, the `addResult` and `addArgument` methods are updated to take
`std::unique_ptr<>`, to match the API for `addInput` and `addOutput`.
The unique pointer is released on the inside, but I like how the unique
pointer makes it crystal clear that ownership is transferred. It also
calms down CLion warnings about possibly leaking memory.