Make TopDown and BottomUp traversers generic for constness (#1335)
This is done by creating two generic Observers. One of them is an actual
observer that forwards all events to the traverser, while the other is a
dummy class that does nothing except take a `const Region &`.

It still uses the same `TraversalTracker` as before, just makes it
generic over the constness of the `Node` type.

It would also be possible to not make the `TraversalTracker` generic,
and instead do `const_cast` every time we need a non-const Node.

Lastly it adds a check that only visited nodes are removed in the bottom
up traverser. Otherwise it can cause bugs.

The new const traversers are called `TopDownConstTraverser` and
`BottomUpConstTraverser`.

FYI @caleridas