Move output tracing and constant integer evaluation out of AliasAnalysis (#1337)
Adds `rvsdg::traceOutput` which expands on
`rvsdg::traceOutputIntraProcedurally` to also trace out of lambda, phi
and delta context variables.

Adds `llvm::traceOutput` which expands on `rvsdg::traceOutput` to also
trace through `IOBarrier`.

Adds `llvm::tryGetConstantSignedInteger`, which traces the given output
to find out if it contains a constant integer value, and if so, what
that value is.

The latter two functions replace behavior that was previously
implemented in the `AliasAnalysis`, because that was where they were
used.