Fail CI on changed cycle count
jlm is now determenistic, so this PR enables the CI to fail if the
simulated cycle count change for the HLS test suite.
A new argument is provided for the run-hls-test.sh script that
updates the golden cycles, which is stored in .github/golden for
easiness of updating them when a PR affects the cycle count.