Implement chain recurrence folding (#1362)
This PR implements logic for creating chain recurrences from SCEV trees.
It also refactors much of the implementation by removing the first
fixed-point analysis and the method to replace placeholders.

While SCEV trees can be nested, chain recurrences are one level only,
and allows us to more easily reason about the step in an induction
variable, something that will be useful when this is applied in an
optimization.

The reason for removing the fixed-point analysis is this: The
fixed-point analysis would traverse the graph and check, for all
induction variable candidates, if any illegal operations was a part of
the tree, and if so, remove them from the candidate set, and we would
only create SCEV trees for the valid candidates. However, we would still
have to check for dependencies between these candidates, and they could
still be invalid as induction variables, so we would have to filter them
twice. I think a safer approach is to create SCEV trees for all loop
variables, and just use SCEVUnknown for illegal operations. This means
that for any output, we can still look up it's SCEV, which gives us
extra information to reason about, that otherwise wouldn't be there. It
is also more efficient.

---------

Co-authored-by: Nico Reissmann <nico.reissmann@gmail.com>