Add new RegionAwareModRefSummarizer, change ModRefSummary interface (#1288)
This PR changes the RegionAwareModRefSummarizer to be smart about
allocas, both in terms of SCCs and Non-Reentrant allocas.

It also changes the interface for `ModRefSummary` to make it produce
more direct info for the `MemoryStateEncoder`, which in turn required
changes to the `AgnosticModRefSummarizer`.

The PR removes the `TopDownModRefEliminator`, as it had correctness
issues, and would be painful to convert to the new interface. (could be
its own PR)

It also fixes an issue with allocas in the `MemoryStateEncoder`, which
would previously be missing join operations. (could be its own PR)

All these changes brought with them a ton of changes to the tests, which
has been taking quite a while (I don't want to just change the `asserts`
in tests without actually checking that the new `asserts` actually make
sense).

When making this PR, I think the only remaining tests that still need to
be updated are in `RegionAwareModRefSummarizerTests.cpp`. I make this
draft PR to let you start looking at it tomorrow morning. I know it is
huge.

---------

Co-authored-by: Helge Bahmann <hcb@chaoticmind.net>