Scalar evolution statistics (#1387)
This is a first version of the scalar evolution statistics
implementation.

I used a companion Context class, and added 6 measurements with
corresponding labels. I am a bit unsure if this is the best way to go
about it, so feedback is greatly appreciated.

This PR also extends the implementation to support for- and while-loops
by utilizing the `tryGetConstantSignedInteger` method in `trace.hpp`, as
a more robust way to get the start values for a loop variable. Thanks to
@haved for the suggestion.

---------

Co-authored-by: Nico Reissmann <nico.reissmann@gmail.com>