Make it possible to enable --LoopUnswitching (#1390)
The `jlm-opt` binary would not handle being given `--LoopUnswitching`,
only `--ThetaGammaInversion`.

This PR fixes this, while moving the deprication of
`--ThetaGammaInversion` a step further.
It also uses the `BijectiveMap` instead of having two separate maps
between OptimizationId and the string.

(It also removes ScalarEvolution from -O3, as that is only an analysis
and not a transformation)