Distinguish between store sizes and alloc sizes, like LLVM does (#1399)
Fixes #1398 

In LLVM's `DataLayout`, there is a difference between
[getTypeStoreSize](https://llvm.org/doxygen/classllvm_1_1DataLayout.html#acfcd22eb38dbfe1acbf138754297437a)
and
[getTypeAllocSize](https://llvm.org/doxygen/classllvm_1_1DataLayout.html#aa424fe8f47ba4ff440cf7f7742a1c162).

We almost always want to use the latter, as it behaves like C's
`sizeof`, in that it is rounded up to the type's own alignment.

However, LLVM IR allows performing, e.g., 40 bit loads, which should NOT
be rounded up to an 8 byte load, as it risks reading out of bounds of a
5 byte allocation. The type `i40` thus has a store-size of 5 bytes, and
an alloc size of 8 bytes.