[AndersenRegionAware] Change the PointsToGraph interface and add flags (#1389)
The new interface is based on indices into packed arrays. It loses the
ability to get a list of sources for a given target node, but that was
not really used anywhere.

The added flags are:
 - points to everything that is externally available
 - this memory node is externally available

We still have the external memory node, as it is useful to have a
NodeIndex to represent "all memory that is not represented by any other
memory node". This node can never be an explicit target, however, and
does not have any explicit targets either.