Make traceOutput more powerful and use it in the call classifier (#1426)
The `CallOperation`'s classifier contained output tracing machinery that
did mostly the same as the regular `traceOutput` and friends, but was a
bit more capable, as it would traverse into subregions to try extra hard
to see if outputs are invariant.

This PR makes the general `traceOutput` as powerful, in a way that does
not require state to be stored. The call classifier is changed to make
use of it instead of having its own duplicated logic.

I do think the solution we currently have to handle IOBarrier from LLVM
is a bit bad, as it requires creating a separate `traceOutput` for when
we are inside the LLVM dialect. It would make more sense if operations
had a general way of reporting that some outputs are invariant, but that
is left for later.

Fixes #1423