Reimplement RegionArgument Removal (#1427)
The performance of the sweep phase of DNE is currently quadratic in the
number of arguments/results/inputs/outputs. The reason for this is that
we remove arguments/results/inputs/outputs one at a time and need to
readjust indices for all other remaining
arguments/results/inputs/outputs while doing so. I would like to move
away from the individual removal of arguments/results/inputs/outputs
towards mass removal.

This is the first step in this direction by reimplementing
`Region::RemoveArgumentsWhere()` to not rely on the individual removal
methods. I will adjust the other `Remove*Where()` methods in future PRs,
and then utilize them in the removal functions of the structural nodes.