Inline pass fix (#1432)
Fixes the inlining pass in the following ways:
 - inlining is not performed if the callee has
   - Varargs
   - Calls to `setjmp`
   - Calls to `alloca` with runtime known sizes
   - Calls to `alloca` that are not on the top level

When inlining is performed, it performs the following extra steps:
- Matches up `CallEntryMerge`/`LambdaEntrySplit` and
`LambdaExitMerge`/`CallExitSplit` to route memory states directly, or
completely around the function body if the memory state id is not used
inside the callee.
- All alloca operations in the callee are hoised to the top level of the
caller

In order to do tracing correctly even in the case of functions in phi
nodes, I had to add an additional configuration option to the
`OutputTracer` that controls if tracing is allowed to enter into phi
nodes. We often want to go into phi nodes to find the actual lambda we
are referring to, but this risks tracing descending into sibling
regions, and not just going up through the region hierarchy.

Adds tests of the alloca handling in inlining.