Replace types for testing (#1451)
This PR does the following:
1. Implements a configurable TestType in the jlm::rvsdg namespace
2. Replaces all uses of jlm::tests::ValueType and jlm::tests::StateType
with the new jlm::rvsdg::TestType
3. Removes test-types.hpp and test-types.cpp, which contained the
jlm::tests::ValueType and jlm::tests::StateType

The goal is to get the libjlmtests library completely dissolved as it is
in the way to convert all unit tests to google tests. It contains a
main() definition, which does not go well with google tests.